requirements.txt
asgiref==3.6.0
asttokens==2.0.8
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.52.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
entrypoints==0.4
executing==1.1.0
idna==3.4
iniconfig==1.1.1
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-client==7.3.5
jupyter-core==4.11.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-environ==0.4.54
python3-openid==3.2.0
pytz==2022.7
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
urllib3==1.26.13
wcwidth==0.2.5
